update publickey auth. OpenSSH pubkey.
